METAMODEL-244: Moved to pkg org.apache.metamodel.schema.naming
