+++ core/src/main/java/org/apache/metamodel/schema/builder/DelegatingIntrinsicSwitchColumnNamingStrategy.java
@@ -18,6 +18,10 @@
  */
 package org.apache.metamodel.schema.builder;
 
+import org.apache.metamodel.schema.naming.ColumnNamingContext;
+import org.apache.metamodel.schema.naming.ColumnNamingSession;
+import org.apache.metamodel.schema.naming.ColumnNamingStrategy;
+
 /**
  * A {@link ColumnNamingStrategy} that switches between two other
  * {@link ColumnNamingStrategy} delegates depending on the availability of a
