+++ core/src/main/java/org/apache/metamodel/schema/naming/ColumnNamingContext.java
@@ -15,7 +15,7 @@
  * KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations
  * under the License.
- */package org.apache.metamodel.schema.builder;
+ */package org.apache.metamodel.schema.naming;
 
 import org.apache.metamodel.schema.Table;
