+++ core/src/main/java/org/apache/metamodel/schema/naming/ColumnNamingContextImpl.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.metamodel.schema.builder;
+package org.apache.metamodel.schema.naming;
 
 import org.apache.metamodel.schema.Table;
